Task#4: change files naming
