+++ client/src/features/ErrorBoundary/style.scss
