+++ client/src/features/MovieCard/style.scss
