+++ client/src/features/MoviesListItem/style.scss
