+++ client/src/shared/Footer/component.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import styles from './footer.scss';
+import styles from './style.scss';
 
 const Footer = () => {
   return (
