+++ client/src/shared/Footer/component.test.js
@@ -1,4 +1,4 @@
-import Footer from './Footer';
+import Footer from './component';
 import React from 'react';
 import { mount } from 'enzyme';
