+++ client/src/shared/Footer/index.js
@@ -1,3 +1,3 @@
-import Footer from './Footer';
+import Footer from './component';
 
 export default Footer;
