+++ client/src/shared/Footer/style.scss
