+++ client/src/shared/Navigation/component.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import styles from './navigation.scss';
+import styles from './style.scss';
 
 const Navigation = props => {
   return <div className={styles.navigation}>{props.children}</div>;
