+++ client/src/shared/Navigation/component.test.js
@@ -1,4 +1,4 @@
-import Navigation from './Navigation';
+import Navigation from './component';
 import React from 'react';
 import { mount } from 'enzyme';
