+++ client/src/shared/Navigation/index.js
@@ -1,3 +1,3 @@
-import Navigation from './Navigation';
+import Navigation from './component';
 
 export default Navigation;
