+++ client/src/shared/Toggle/index.js
@@ -1,3 +1,3 @@
-import Toggle from './Toggle';
+import Toggle from './component';
 
 export default Toggle;
