+++ client/src/shared/Toggle/style.scss
